code/django_rest_api/requirements.txt
Django==3.0.7
djangorestframework==3.10.0
djangorestframework-simplejwt==4.6.0
djoser==2.0.3
Pillow==7.2.0
django-cors-headers==3.4.0